models: move Repo and Worker models to own modules (bug 1913168) (#102)
- move Repo model to its own module
- move Worker model to its own module
- update imports
