+++ src/lando/main/models/__init__.py
@@ -4,3 +4,5 @@ from lando.main.models.configuration import *
 from lando.main.models.landing_job import *
 from lando.main.models.profile import *
 from lando.main.models.revision import *
+from lando.main.models.repo import *
+from lando.main.models.worker import *
