+++ src/lando/main/models/worker.py
@@ -1,0 +1,25 @@
+import logging
+import os
+
+from django.db import models
+
+from lando.main.models import BaseModel, Repo
+
+logger = logging.getLogger(__name__)
+
+DEFAULT_GRACE_SECONDS = int(os.environ.get("DEFAULT_GRACE_SECONDS", 60 * 2))
+
+
+class Worker(BaseModel):
+    name = models.CharField(max_length=255, unique=True)
+    is_paused = models.BooleanField(default=False)
+    is_stopped = models.BooleanField(default=False)
+    ssh_private_key = models.TextField(null=True, blank=True)
+    applicable_repos = models.ManyToManyField(Repo)
+
+    throttle_seconds = models.IntegerField(default=10)
+    sleep_seconds = models.IntegerField(default=10)
+
+    @property
+    def enabled_repos(self) -> list[Repo]:
+        return self.applicable_repos.all()
